Move highlightProvider to standalone file
